Update topology-utils, move to topology2
