+++ frontend/packages/dev-console/src/components/topology/TopologyDataController.tsx
@@ -6,8 +6,8 @@ import { getResourceList } from '@console/shared';
 import { referenceForModel } from '@console/internal/module/k8s';
 import { ClusterServiceVersionModel } from '@console/operator-lifecycle-manager/src/models';
 import { ServiceBindingRequestModel } from '../../models';
+import { allowedResources, transformTopologyData } from '../topology2/topology-utils';
 import { TopologyDataModel, TopologyDataResources } from './topology-types';
-import { allowedResources, transformTopologyData } from './topology-utils';
 
 export interface RenderProps {
   data?: TopologyDataModel;
