+++ frontend/packages/dev-console/src/components/topology/actions/nodeActions.ts
@@ -2,7 +2,7 @@ import * as _ from 'lodash';
 import { Kebab, KebabOption } from '@console/internal/components/utils';
 import { modelFor, referenceFor } from '@console/internal/module/k8s';
 import { TopologyDataObject } from '../topology-types';
-import { getTopologyResourceObject } from '../topology-utils';
+import { getTopologyResourceObject } from '../../topology2/topology-utils';
 
 export const nodeActions = (node: TopologyDataObject): KebabOption[] => {
   const contextMenuResource = getTopologyResourceObject(node);
