+++ frontend/packages/dev-console/src/components/topology2/components/createConnection.tsx
@@ -1,0 +1,16 @@
+import { Node } from '@console/topology';
+import { createTopologyResourceConnection } from '../topology-utils';
+
+export const createConnection = (
+  sourceNode: Node,
+  targetNode: Node,
+  replaceTargetNode: Node = null,
+  serviceBindingFlag: boolean,
+): Promise<any> => {
+  return createTopologyResourceConnection(
+    sourceNode.getData(),
+    targetNode.getData(),
+    replaceTargetNode ? replaceTargetNode.getData() : null,
+    serviceBindingFlag,
+  );
+};
