+++ frontend/packages/dev-console/src/components/topology2/components/nodes/KnativeService.tsx
@@ -18,7 +18,7 @@ import {
 } from '@console/topology';
 import { useAccessReview } from '@console/internal/components/utils';
 import { modelFor, referenceFor } from '@console/internal/module/k8s';
-import { getTopologyResourceObject } from '../../../topology/topology-utils';
+import { getTopologyResourceObject } from '../../topology-utils';
 import SvgBoxedText from '../../../svg/SvgBoxedText';
 import Decorator from '../../../topology/shapes/Decorator';
 import RevisionTrafficSourceAnchor from '../anchors/RevisionTrafficSourceAnchor';
